requirements.txt
numpy
pandas 
python-dotenv
mysql-connector-python
pymysql
scikit-learn
seaborn
catboost
xgboost
mlflow==2.0.0
pyarrow==9.0.0
flask
#-e .